Move isExpressionDependentOnThis to ExpressionDependsOnThisHelper
Extract the static method into a standalone helper class to avoid
ConstantConditionRuleHelper depending on ImpossibleCheckTypeHelper.
